Add rvm_io.ruby as a dependency of the jekyll-build role
The goal of this dependency is to prevent being tied to the Ruby version
installed by the package manager (currently 2.0) and use rvm instead. The main
downside is the requirement for which to be installed on EL/CentOS. This needs
to be performed as a pre-task at the moment although we could consider
alternate strategies.